Restructured the website by grouping all the user-side pages into /home
