+++ app/home/book-appointment/page.tsx
@@ -1,0 +1,14 @@
+export default function BookAppointmentPage() {
+  return (
+    <div>
+      <section className="bg-gray-100 py-16">
+        <div className="container mx-auto text-center">
+          <h2 className="text-4xl font-bold text-gray-800">Book Appointment</h2>
+          <p className="mt-4 text-lg text-gray-600 max-w-2xl mx-auto">
+            Book an appointment with one of our doctors.
+          </p>
+        </div>
+      </section>
+    </div>
+  );
+}
